+++ packages/js/product-editor/src/blocks/index.ts
@@ -27,4 +27,4 @@ export { init as initVariationItems } from './variation-items';
 export { init as initVariationOptions } from './variation-options';
 export { init as initNotice } from './notice';
 export { init as initTaxonomy } from './taxonomy';
-export { init as initInput } from './input';
+export { init as initInput } from './text';
