+++ packages/js/product-editor/src/blocks/style.scss
@@ -19,4 +19,4 @@
 @import 'variation-options/editor.scss';
 @import 'taxonomy/editor.scss';
 @import 'toggle/editor.scss';
-@import 'input/editor.scss';
+@import 'text/editor.scss';
