+++ packages/js/product-editor/src/blocks/text/block.json
