+++ packages/js/product-editor/src/blocks/text/index.ts
